backend/app/utils/auth.py
import os
from datetime import datetime, timedelta, timezone

from fastapi import Depends, Header, HTTPException, status
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy import select

from app.models import User
from app.db import get_db

from dotenv import load_dotenv
import jwt


load_dotenv()  # load .env

SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = os.getenv("ALGORITHM")


def verify_access_token(
    authorization: str = Header(...)) -> dict:
    """
    從 Header 拿 Bearer token，驗證合法性並解出 payload。
    """
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    # 1. Authorization header 
    if not authorization.startswith("Bearer "):
        raise credentials_exception
    token = authorization.split(" ")[1]

    # 2. Decode the JWT token
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        raise credentials_exception

# Get the current user from the token payload
async def get_current_user(
    payload: dict = Depends(verify_access_token),
    db: AsyncSession = Depends(get_db)
):
    email = payload.get("sub")
    stmt = select(User).where(User.email == email)
    result = await db.execute(stmt)
    user = result.scalars().first()
    return user


def create_access_token(data: dict, expires_delta: timedelta = timedelta(hours=1)):
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + expires_delta
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
