backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["shuxinqiao <sqiao@ualberta.ca>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.13"
uvicorn = {extras = ["standard"], version = "^0.34.3"}
email-validator = "^2.2.0"
sqlalchemy = "^2.0.41"
python-multipart = "^0.0.20"
passlib = "^1.7.4"
aiosqlite = "^0.21.0"
pyjwt = "^2.10.1"
bcrypt = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
